+++ src/CLaSH/Prelude/Explicit.hs
@@ -47,6 +47,7 @@ module CLaSH.Prelude.Explicit
   , isRising'
   , isFalling'
     -- * Testbench functions
+  , assert'
   , stimuliGenerator'
   , outputVerifier'
     -- * Exported modules
@@ -55,48 +56,23 @@ module CLaSH.Prelude.Explicit
   )
 where
 
-import Control.Applicative     (liftA2)
-import Data.Default            (Default (..))
-import GHC.TypeLits            (KnownNat, type (+), natVal)
-import Prelude                 hiding (repeat)
+import Data.Default                 (Default (..))
+import GHC.TypeLits                 (KnownNat, type (+), natVal)
+import Prelude                      hiding (repeat)
 
-import CLaSH.Prelude.BlockRam  (blockRam', blockRamPow2')
+import CLaSH.Prelude.Explicit.Safe
 import CLaSH.Prelude.BlockRam.File (blockRamFile', blockRamFilePow2')
-import CLaSH.Prelude.Mealy     (mealy', mealyB')
-import CLaSH.Prelude.Moore     (moore', mooreB')
-import CLaSH.Prelude.RAM       (asyncRam',asyncRamPow2')
-import CLaSH.Prelude.ROM       (rom', romPow2')
-import CLaSH.Prelude.ROM.File  (romFile', romFilePow2')
-import CLaSH.Prelude.Testbench (stimuliGenerator', outputVerifier')
+import CLaSH.Prelude.ROM.File      (romFile', romFilePow2')
+import CLaSH.Prelude.Testbench     (assert', stimuliGenerator', outputVerifier')
 import CLaSH.Signal.Explicit
-import CLaSH.Sized.Vector      (Vec (..), (+>>), asNatProxy, repeat)
+import CLaSH.Sized.Vector          (Vec (..), (+>>), asNatProxy, repeat)
 
 -- $setup
 -- >>> :set -XDataKinds
 -- >>> type ClkA = Clk "A" 100
 -- >>> let clkA = sclock :: SClock ClkA
--- >>> let rP = registerB' clkA (8::Int,8::Int)
 -- >>> let window4 = window' clkA :: Signal' ClkA Int -> Vec 4 (Signal' ClkA Int)
 -- >>> let windowD3 = windowD' clkA :: Signal' ClkA Int -> Vec 3 (Signal' ClkA Int)
-
-{-# INLINE registerB' #-}
--- | Create a 'register' function for product-type like signals (e.g.
--- @('Signal' a, 'Signal' b)@)
---
--- @
--- type ClkA = 'Clk' \"A\" 100
---
--- clkA :: 'SClock' ClkA
--- clkA = 'sclock'
---
--- rP :: ('Signal'' ClkA Int, 'Signal'' ClkA Int) -> ('Signal'' ClkA Int, 'Signal'' ClkA Int)
--- rP = 'registerB'' clkA (8,8)
--- @
---
--- >>> simulateB' clkA clkA rP [(1,1),(2,2),(3,3)] :: [(Int,Int)]
--- [(8,8),(1,1),(2,2),(3,3)...
-registerB' :: Bundle a => SClock clk -> a -> Unbundled' clk a -> Unbundled' clk a
-registerB' clk i = unbundle' clk Prelude.. register' clk i Prelude.. bundle' clk
 
 {-# INLINABLE window' #-}
 -- | Give a window over a 'Signal''
@@ -150,27 +126,3 @@ windowD' clk x = prev
   where
     prev = registerB' clk (repeat def) next
     next = x +>> prev
-
-{-# INLINABLE isRising' #-}
--- | Give a pulse when the 'Signal'' goes from 'minBound' to 'maxBound'
-isRising' :: (Bounded a, Eq a)
-          => SClock clk
-          -> a -- ^ Starting value
-          -> Signal' clk a
-          -> Signal' clk Bool
-isRising' clk is s = liftA2 edgeDetect prev s
-  where
-    prev = register' clk is s
-    edgeDetect old new = old == minBound && new == maxBound
-
-{-# INLINABLE isFalling' #-}
--- | Give a pulse when the 'Signal'' goes from 'maxBound' to 'minBound'
-isFalling' :: (Bounded a, Eq a)
-           => SClock clk
-           -> a -- ^ Starting value
-           -> Signal' clk a
-           -> Signal' clk Bool
-isFalling' clk is s = liftA2 edgeDetect prev s
-  where
-    prev = register' clk is s
-    edgeDetect old new = old == maxBound && new == minBound